Divide box_types.go by specification #139
