+++ box_types_iso14496_12_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestBoxTypes(t *testing.T) {
+func TestBoxTypesISO14496_12(t *testing.T) {
 	testCases := []struct {
 		name string
 		src  IImmutableBox
@@ -373,107 +373,6 @@ func TestBoxTypes(t *testing.T) {
 				`MessageData="abema"`,
 		},
 		{
-			name: "esds",
-			src: &Esds{
-				FullBox: FullBox{
-					Version: 0,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				Descriptors: []Descriptor{
-					{
-						Tag:  ESDescrTag,
-						Size: 0x1234567,
-						ESDescriptor: &ESDescriptor{
-							ESID:                 0x1234,
-							StreamDependenceFlag: true,
-							UrlFlag:              false,
-							OcrStreamFlag:        true,
-							StreamPriority:       0x03,
-							DependsOnESID:        0x2345,
-							OCRESID:              0x3456,
-						},
-					},
-					{
-						Tag:  ESDescrTag,
-						Size: 0x1234567,
-						ESDescriptor: &ESDescriptor{
-							ESID:                 0x1234,
-							StreamDependenceFlag: false,
-							UrlFlag:              true,
-							OcrStreamFlag:        false,
-							StreamPriority:       0x03,
-							URLLength:            11,
-							URLString:            []byte("http://hoge"),
-						},
-					},
-					{
-						Tag:  DecoderConfigDescrTag,
-						Size: 0x1234567,
-						DecoderConfigDescriptor: &DecoderConfigDescriptor{
-							ObjectTypeIndication: 0x12,
-							StreamType:           0x15,
-							UpStream:             true,
-							Reserved:             false,
-							BufferSizeDB:         0x123456,
-							MaxBitrate:           0x12345678,
-							AvgBitrate:           0x23456789,
-						},
-					},
-					{
-						Tag:  DecSpecificInfoTag,
-						Size: 0x03,
-						Data: []byte{0x11, 0x22, 0x33},
-					},
-					{
-						Tag:  SLConfigDescrTag,
-						Size: 0x05,
-						Data: []byte{0x11, 0x22, 0x33, 0x44, 0x55},
-					},
-				},
-			},
-			dst: &Esds{},
-			bin: []byte{
-				0,                // version
-				0x00, 0x00, 0x00, // flags
-				//
-				0x03,                   // tag
-				0x89, 0x8d, 0x8a, 0x67, // size (varint)
-				0x12, 0x34, // esid
-				0xa3,       // flags & streamPriority
-				0x23, 0x45, // dependsOnESID
-				0x34, 0x56, // ocresid
-				//
-				0x03,                   // tag
-				0x89, 0x8d, 0x8a, 0x67, // size (varint)
-				0x12, 0x34, // esid
-				0x43,                                                  // flags & streamPriority
-				11,                                                    // urlLength
-				'h', 't', 't', 'p', ':', '/', '/', 'h', 'o', 'g', 'e', // urlString
-				//
-				0x04,                   // tag
-				0x89, 0x8d, 0x8a, 0x67, // size (varint)
-				0x12,             // objectTypeIndication
-				0x56,             // streamType & upStream & reserved
-				0x12, 0x34, 0x56, // bufferSizeDB
-				0x12, 0x34, 0x56, 0x78, // maxBitrate
-				0x23, 0x45, 0x67, 0x89, // avgBitrate
-				//
-				0x05,                   // tag
-				0x80, 0x80, 0x80, 0x03, // size (varint)
-				0x11, 0x22, 0x33, // data
-				//
-				0x06,                   // tag
-				0x80, 0x80, 0x80, 0x05, // size (varint)
-				0x11, 0x22, 0x33, 0x44, 0x55, // data
-			},
-			str: `Version=0 Flags=0x000000 Descriptors=[` +
-				`{Tag=ESDescr Size=19088743 ESID=4660 StreamDependenceFlag=true UrlFlag=false OcrStreamFlag=true StreamPriority=3 DependsOnESID=9029 OCRESID=13398}, ` +
-				`{Tag=ESDescr Size=19088743 ESID=4660 StreamDependenceFlag=false UrlFlag=true OcrStreamFlag=false StreamPriority=3 URLLength=0xb URLString="http://hoge"}, ` +
-				`{Tag=DecoderConfigDescr Size=19088743 ObjectTypeIndication=0x12 StreamType=21 UpStream=true Reserved=false BufferSizeDB=1193046 MaxBitrate=305419896 AvgBitrate=591751049}, ` +
-				"{Tag=DecSpecificInfo Size=3 Data=[0x11, 0x22, 0x33]}, " +
-				"{Tag=SLConfigDescr Size=5 Data=[0x11, 0x22, 0x33, 0x44, 0x55]}]",
-		},
-		{
 			name: "fiel",
 			src: &Fiel{
 				FieldCount:    233,
@@ -664,148 +563,6 @@ func TestBoxTypes(t *testing.T) {
 				`NaluType=0x22 NumNalus=1 Nalus=[{Length=7 NALUnit=[0x44, 0x1, 0xc1, 0x72, 0xb4, 0x62, 0x40]}]}, ` +
 				`{Completeness=false Reserved=false NaluType=0x27 NumNalus=1 Nalus=[{Length=11 NALUnit=[0x4e, 0x1, 0x5, 0xff, 0xff, 0xff, ` +
 				`0xa6, 0x2c, 0xa2, 0xde, 0x9]}]}]`,
-		},
-		{
-			name: "ilst",
-			src:  &Ilst{},
-			dst:  &Ilst{},
-			bin:  nil,
-			str:  ``,
-		},
-		{
-			name: "ilst meta container",
-			src: &IlstMetaContainer{
-				AnyTypeBox: AnyTypeBox{Type: StrToBoxType("----")},
-			},
-			dst: &IlstMetaContainer{
-				AnyTypeBox: AnyTypeBox{Type: StrToBoxType("----")},
-			},
-			bin: nil,
-			str: ``,
-			ctx: Context{UnderIlst: true},
-		},
-		{
-			name: "ilst data (binary)",
-			src:  &Data{DataType: 0, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x00, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=BINARY DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (utf8)",
-			src:  &Data{DataType: 1, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x01, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=UTF8 DataLang=305419896 Data="foo"`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (utf8 escaped)",
-			src:  &Data{DataType: 1, DataLang: 0x12345678, Data: []byte{0x00, 'f', 'o', 'o'}},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x01, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x00, 0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=UTF8 DataLang=305419896 Data=".foo"`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (utf16)",
-			src:  &Data{DataType: 2, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x02, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=UTF16 DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (mac string)",
-			src:  &Data{DataType: 3, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x03, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=MAC_STR DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (jpsg)",
-			src:  &Data{DataType: 14, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x0e, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=JPEG DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (int)",
-			src:  &Data{DataType: 21, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x15, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=INT DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (float32)",
-			src:  &Data{DataType: 22, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x16, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=FLOAT32 DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (float64)",
-			src:  &Data{DataType: 23, DataLang: 0x12345678, Data: []byte("foo")},
-			dst:  &Data{},
-			bin: []byte{
-				0x00, 0x00, 0x00, 0x17, // data type
-				0x12, 0x34, 0x56, 0x78, // data lang
-				0x66, 0x6f, 0x6f, // data
-			},
-			str: `DataType=FLOAT64 DataLang=305419896 Data=[0x66, 0x6f, 0x6f]`,
-			ctx: Context{UnderIlstMeta: true},
-		},
-		{
-			name: "ilst data (string)",
-			src: &StringData{
-				AnyTypeBox: AnyTypeBox{Type: StrToBoxType("mean")},
-				Data:       []byte{0x00, 'f', 'o', 'o'},
-			},
-			dst: &StringData{
-				AnyTypeBox: AnyTypeBox{Type: StrToBoxType("mean")},
-			},
-			bin: []byte{
-				0x00, 0x66, 0x6f, 0x6f, // data
-			},
-			str: `Data=".foo"`,
-			ctx: Context{UnderIlstFreeMeta: true},
 		},
 		{
 			name: "mdat",
@@ -1104,67 +861,6 @@ func TestBoxTypes(t *testing.T) {
 				`Matrix=[0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0] ` +
 				`PreDefined=[0, 0, 0, 0, 0, 0] ` +
 				`NextTrackID=2882400001`,
-		},
-		{
-			name: "pssh: version 0: no KIDs",
-			src: &Pssh{
-				FullBox: FullBox{
-					Version: 0,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				SystemID: [16]byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10},
-				DataSize: 5,
-				Data:     []byte{0x21, 0x22, 0x23, 0x24, 0x25},
-			},
-			dst: &Pssh{},
-			bin: []byte{
-				0,                // version
-				0x00, 0x00, 0x00, // flags
-				// system ID
-				0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10,
-				0x00, 0x00, 0x00, 0x05, // data size
-				0x21, 0x22, 0x23, 0x24, 0x25, // data
-			},
-			str: `Version=0 Flags=0x000000 ` +
-				`SystemID=01020304-0506-0708-090a-0b0c0d0e0f10 ` +
-				`DataSize=5 ` +
-				`Data=[0x21, 0x22, 0x23, 0x24, 0x25]`,
-		},
-		{
-			name: "pssh: version 1: with KIDs",
-			src: &Pssh{
-				FullBox: FullBox{
-					Version: 1,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				SystemID: [16]byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10},
-				KIDCount: 2,
-				KIDs: []PsshKID{
-					{KID: [16]byte{0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f, 0x10}},
-					{KID: [16]byte{0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f, 0x20}},
-				},
-				DataSize: 5,
-				Data:     []byte{0x21, 0x22, 0x23, 0x24, 0x25},
-			},
-			dst: &Pssh{},
-			bin: []byte{
-				1,                // version
-				0x00, 0x00, 0x00, // flags
-				// system ID
-				0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10,
-				0x00, 0x00, 0x00, 0x02, // KID count
-				// KIDs
-				0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f, 0x10,
-				0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f, 0x20,
-				0x00, 0x00, 0x00, 0x05, // data size
-				0x21, 0x22, 0x23, 0x24, 0x25, // data
-			},
-			str: `Version=1 Flags=0x000000 ` +
-				`SystemID=01020304-0506-0708-090a-0b0c0d0e0f10 ` +
-				`KIDCount=2 ` +
-				`KIDs=[11121314-1516-1718-191a-1b1c1d1e1f10, 21222324-2526-2728-292a-2b2c2d2e2f20] ` +
-				`DataSize=5 ` +
-				`Data=[0x21, 0x22, 0x23, 0x24, 0x25]`,
 		},
 		{
 			name: "saio: version 0: no aux info type",
@@ -2519,117 +2215,6 @@ func TestBoxTypes(t *testing.T) {
 			str: `MajorBrand="abem" MinorVersion=305419896 CompatibleBrands=[{CompatibleBrand="abcd"}, {CompatibleBrand="efgh"}]`,
 		},
 		{
-			name: "tenc: DefaultIsProtected=1 DefaultPerSampleIVSize=0",
-			src: &Tenc{
-				FullBox: FullBox{
-					Version: 1,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				Reserved:               0x00,
-				DefaultCryptByteBlock:  0x0a,
-				DefaultSkipByteBlock:   0x0b,
-				DefaultIsProtected:     1,
-				DefaultPerSampleIVSize: 0,
-				DefaultKID: [16]byte{
-					0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-					0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-				},
-				DefaultConstantIVSize: 4,
-				DefaultConstantIV:     []byte{0x01, 0x23, 0x45, 0x67},
-			},
-			dst: &Tenc{},
-			bin: []byte{
-				1,                // version
-				0x00, 0x00, 0x00, // flags
-				0x00,       // reserved
-				0xab,       // default crypt byte block / default skip byte block
-				0x01, 0x00, // default is protected / default per sample iv size
-				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef, // default kid
-				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-				0x04,                   // default constant iv size
-				0x01, 0x23, 0x45, 0x67, // default constant iv
-			},
-			str: `Version=1 Flags=0x000000 ` +
-				`Reserved=0 ` +
-				`DefaultCryptByteBlock=10 ` +
-				`DefaultSkipByteBlock=11 ` +
-				`DefaultIsProtected=1 ` +
-				`DefaultPerSampleIVSize=0 ` +
-				`DefaultKID=01234567-89ab-cdef-0123-456789abcdef ` +
-				`DefaultConstantIVSize=4 ` +
-				`DefaultConstantIV=[0x1, 0x23, 0x45, 0x67]`,
-		},
-		{
-			name: "tenc: DefaultIsProtected=0 DefaultPerSampleIVSize=0",
-			src: &Tenc{
-				FullBox: FullBox{
-					Version: 1,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				Reserved:               0x00,
-				DefaultCryptByteBlock:  0x0a,
-				DefaultSkipByteBlock:   0x0b,
-				DefaultIsProtected:     0,
-				DefaultPerSampleIVSize: 0,
-				DefaultKID: [16]byte{
-					0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-					0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-				},
-			},
-			dst: &Tenc{},
-			bin: []byte{
-				1,                // version
-				0x00, 0x00, 0x00, // flags
-				0x00,       // reserved
-				0xab,       // default crypt byte block / default skip byte block
-				0x00, 0x00, // default is protected / default per sample iv size
-				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef, // default kid
-				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-			},
-			str: `Version=1 Flags=0x000000 ` +
-				`Reserved=0 ` +
-				`DefaultCryptByteBlock=10 ` +
-				`DefaultSkipByteBlock=11 ` +
-				`DefaultIsProtected=0 ` +
-				`DefaultPerSampleIVSize=0 ` +
-				`DefaultKID=01234567-89ab-cdef-0123-456789abcdef`,
-		},
-		{
-			name: "tenc: DefaultIsProtected=1 DefaultPerSampleIVSize=1",
-			src: &Tenc{
-				FullBox: FullBox{
-					Version: 1,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				Reserved:               0x00,
-				DefaultCryptByteBlock:  0x0a,
-				DefaultSkipByteBlock:   0x0b,
-				DefaultIsProtected:     1,
-				DefaultPerSampleIVSize: 1,
-				DefaultKID: [16]byte{
-					0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-					0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-				},
-			},
-			dst: &Tenc{},
-			bin: []byte{
-				1,                // version
-				0x00, 0x00, 0x00, // flags
-				0x00,       // reserved
-				0xab,       // default crypt byte block / default skip byte block
-				0x01, 0x01, // default is protected / default per sample iv size
-				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef, // default kid
-				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
-			},
-			str: `Version=1 Flags=0x000000 ` +
-				`Reserved=0 ` +
-				`DefaultCryptByteBlock=10 ` +
-				`DefaultSkipByteBlock=11 ` +
-				`DefaultIsProtected=1 ` +
-				`DefaultPerSampleIVSize=1 ` +
-				`DefaultKID=01234567-89ab-cdef-0123-456789abcdef`,
-		},
-		{
 			name: "tfdt: version 0",
 			src: &Tfdt{
 				FullBox: FullBox{
@@ -3107,25 +2692,6 @@ func TestBoxTypes(t *testing.T) {
 			dst:  &Udta{},
 			bin:  nil,
 			str:  ``,
-		},
-		{
-			name: "udta 3gpp string",
-			src: &Udta3GppString{
-				AnyTypeBox: AnyTypeBox{Type: StrToBoxType("titl")},
-				Language:   [3]byte{0x5, 0xe, 0x7},
-				Data:       []byte("SING"),
-			},
-			dst: &Udta3GppString{
-				AnyTypeBox: AnyTypeBox{Type: StrToBoxType("titl")},
-			},
-			bin: []byte{
-				0,                // version
-				0x00, 0x00, 0x00, // flags
-				0x15, 0xc7, // language
-				0x53, 0x49, 0x4e, 0x47, // data
-			},
-			str: `Version=0 Flags=0x000000 Language="eng" Data="SING"`,
-			ctx: Context{UnderUdta: true},
 		},
 		{
 			name: "vmhd",
